Introduced support for a parameter "--files" for manual tests. [skip ci]
