+++ packages/ckeditor5-dev-tests/lib/utils/manual-tests/compilescripts.js
@@ -7,14 +7,23 @@
 
 'use strict';
 
-const path = require( 'path' );
 const webpack = require( 'webpack' );
 const globSync = require( '../glob' );
 const getWebpackConfigForManualTests = require( './getwebpackconfig' );
 const getRelativeFilePath = require( '../getrelativefilepath' );
 
-module.exports = function compileManualTestScripts( buildDir, manualTestPattern ) {
-	const entryFiles = globSync( path.join( manualTestPattern, '*.js' ) );
+/**
+ * @param {String} buildDir A path where compiled files will be saved.
+ * @param {Array.<String>} manualTestScriptsPattern An array of patterns that resolve manual test scripts.
+ * @returns {Promise}
+ */
+module.exports = function compileManualTestScripts( buildDir, manualTestScriptsPattern ) {
+	const entryFiles = manualTestScriptsPattern.reduce( ( arr, manualTestPattern ) => {
+		arr.push( ...globSync( manualTestPattern ) );
+
+		return arr;
+	}, [] );
+
 	const entries = getWebpackEntryPoints( entryFiles );
 	const webpackConfig = getWebpackConfigForManualTests( entries, buildDir );
